Implement core app structure with models, theme, and swipe interface
