+++ property_swipe/lib/features/swipe/presentation/widgets/property_card.dart
