+++ property_swipe/lib/models/property_listing.dart
