+++ property_swipe/lib/models/user_profile.dart
@@ -239,5 +239,5 @@ enum SwipeAction {
   @JsonValue('right')
   right,
   @JsonValue('super')
-  super,
+  superLike,
 }
