+++ property_swipe/pubspec.yaml
@@ -66,6 +66,9 @@ dependencies:
   # UI Enhancements
   shimmer: ^3.0.0
   lottie: ^2.7.0
+  
+  # JSON Serialization
+  json_annotation: ^4.8.1
 
 dev_dependencies:
   flutter_test:
